Ottimizzati tipi di dati per chiavi primarie

+++ database/migrations/2022_11_06_210531_create_flames_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('flames', function (Blueprint $table) {
-            $table->unsignedSmallInteger('id')->autoIncrement();
+            $table->unsignedTinyInteger('id')->autoIncrement();
             $table->string('name');
             $table->timestamps();
         });
