+++ database/migrations/2022_12_06_155921_create_crosses_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('crosses', function (Blueprint $table) {
-            $table->unsignedTinyInteger('id')->autoIncrement();
+            $table->unsignedSmallInteger('id')->autoIncrement();
             $table->string('name');
             $table->timestamps();
         });
